Reorg to make way for further changes, such as separating training from evaluation.
