+++ scripts/run_question_coding.sh
@@ -2,7 +2,7 @@
 source ~/venv/iep/bin/activate
 
 #ROOT_DIR="/coc/scratch/rvedantam3/runs/pytorch_discovery/question_coding"
-ROOT_DIR="/coc/scratch/rvedantam3/runs/pytorch_discovery/question_coding_num_workers_3"
+ROOT_DIR="/coc/scratch/rvedantam3/runs/pytorch_discovery/question_coding_num_workers_1_load_train"
 PRIOR_CHECKPOINT="/coc/scratch/rvedantam3/runs/pytorch_discovery/prior/lr_0.001/prior.pth"
 DEBUG=0
 
@@ -55,7 +55,7 @@ else
   
     CMD_STRING="python scripts/train_model.py \
       --model_type PG\
-      --loader_num_workers 3\
+      --loader_num_workers 1\
       --model_version $VERSION \
       --program_prior_start_from ${PRIOR_CHECKPOINT} \
       --program_supervision_npy /srv/share/datasets/clevr/CLEVR_v1.0/clevr-iep-data/semi_supervised_train_$supervision.npy \
@@ -65,7 +65,7 @@ else
       --learning_rate 5e-4\
       --checkpoint_path ${RUN_TRAIN_DIR}/question_coding.pth"
 
-    if [ $supervision -eq "699989" ]; then
+    if [ $supervision -gt "1000" ]; then
       CMD_STRING=${CMD_STRING}" --load_train_features_memory"
     fi
   
@@ -78,11 +78,12 @@ else
 
   SSL_ALPHA=100.0
   KL_BETA=0.1
-  MIXING_FACTOR=0.5
+  MIXING_FACTOR=1.0
 
   for supervision in 100 500 5000 10000 699989
+  #for supervision in  500 
   do
-    JOB_STRING="$VERSION"_"${supervision}"
+    JOB_STRING="$VERSION"_"${supervision}"_mix_"$MIXING_FACTOR"_ssl_"${SSL_ALPHA}"
     RUN_TRAIN_DIR=${ROOT_DIR}/${JOB_STRING}
     TRAINVAL_STRING="question_coding"
   
@@ -98,7 +99,7 @@ else
   
     CMD_STRING="python scripts/train_model.py \
       --model_type PG\
-      --loader_num_workers 3\
+      --loader_num_workers 1\
       --model_version $VERSION \
       --program_prior_start_from ${PRIOR_CHECKPOINT} \
       --discovery_alpha ${SSL_ALPHA} \
@@ -110,11 +111,53 @@ else
       --learning_rate 5e-4\
       --checkpoint_path ${RUN_TRAIN_DIR}/question_coding.pth"
 
-    if [ $supervision -gt "5000" ]; then
+    if [ $supervision -gt "100" ]; then
       CMD_STRING=${CMD_STRING}" --load_train_features_memory"
     fi
   
     #exec ${CMD_STRING}
     source utils/invoke_slurm.sh "Y" "${CMD_STRING}" "${JOB_STRING}" "${TRAINVAL_STRING}" "${RUN_TRAIN_DIR}"
+  done
+
+  # MORE DETAILED RUNS FOR 500
+  for supervision in 500
+  do
+    for SSL_ALPHA in 10.0 200.0 1000.0
+    do
+      JOB_STRING="$VERSION"_"${supervision}"_mix_"$MIXING_FACTOR"_ssl_"${SSL_ALPHA}"
+      RUN_TRAIN_DIR=${ROOT_DIR}/${JOB_STRING}
+      TRAINVAL_STRING="question_coding"
+  
+      if [ ! -e ${RUN_TRAIN_DIR} ]; then
+        mkdir ${RUN_TRAIN_DIR}
+      fi
+
+      if [ $supervision -lt 1000 ]; then
+        CHECKPOINT_EVERY=100
+      else
+        CHECKPOINT_EVERY=1000
+      fi
+  
+      CMD_STRING="python scripts/train_model.py \
+        --model_type PG\
+        --loader_num_workers 1\
+        --model_version $VERSION \
+        --program_prior_start_from ${PRIOR_CHECKPOINT} \
+        --discovery_alpha ${SSL_ALPHA} \
+        --discovery_beta ${KL_BETA} \
+        --program_supervision_npy /srv/share/datasets/clevr/CLEVR_v1.0/clevr-iep-data/semi_supervised_train_$supervision.npy \
+        --num_iterations 20000 \
+        --checkpoint_every ${CHECKPOINT_EVERY} \
+        --mixing_factor_supervision ${MIXING_FACTOR} \
+        --learning_rate 5e-4\
+        --checkpoint_path ${RUN_TRAIN_DIR}/question_coding.pth"
+
+      if [ $supervision -gt "100" ]; then
+        CMD_STRING=${CMD_STRING}" --load_train_features_memory"
+      fi
+  
+      #exec ${CMD_STRING}
+      source utils/invoke_slurm.sh "Y" "${CMD_STRING}" "${JOB_STRING}" "${TRAINVAL_STRING}" "${RUN_TRAIN_DIR}"
+    done
   done
 fi
