+++ utils/slurm_wrapper.sh
@@ -6,7 +6,7 @@ sbatch<<EOT
 #SBATCH --output=${4}/${3}.log
 #SBATCH --nodes=1
 #SBATCH --gres=gpu:1
-#SBATCH --exclude=hal
+#SBATCH --exclude=hal,breq
 
 echo "Starting job "${2}
 echo "Running command" ${1}
